+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- In den Rechnungseinstellungen kann jetzt für das Logo unten links die Option "Logo auf allen Seiten anzeigen" aktiviert werden (hitobito_sww#279)
 - In den Rechnungseinstellungen kann jetzt eine Kopfzeile konfiguriert werden (hitobito_sww#278)
 
 ## Version 2.6